requirements.txt
googletrans==3.1.0a0
python-dotenv==1.0.1
requests==2.31.0
google-generativeai==0.8.4